+++ s/core/endpoint.ts
@@ -15,7 +15,7 @@ export type EndpointOptions<F extends Fns> = {
  *  - for each request, it calls the appropriate fn
  *  - it then returns the fn's in json rpc response format
  */
-export function endpoint<F extends Fns>({fns, tap = defaultTap}: EndpointOptions<F>): Endpoint {
+export function makeEndpoint<F extends Fns>({fns, tap = defaultTap}: EndpointOptions<F>): Endpoint {
 	return async request => {
 		const path = request.method.split(".")
 		const fn = drill(fns, path) as Fn