+++ s/transports/http/remote.ts
@@ -1,5 +1,5 @@
 
-import {remote} from "../../core/remote.js"
+import {makeRemote} from "../../core/remote.js"
 import {defaultTap} from "../../core/taps/default.js"
 import {Endpoint, Fns, Tap} from "../../core/types.js"
 
@@ -9,7 +9,7 @@ export type HttpEndpointOptions = {
 }
 
 export function httpRemote<F extends Fns>(options: HttpEndpointOptions) {
-	return remote<F>({endpoint: httpEndpoint(options)})
+	return makeRemote<F>({endpoint: httpEndpoint(options)})
 }
 
 export function httpEndpoint({url, tap = defaultTap}: HttpEndpointOptions): Endpoint {