+++ s/transports/websocket/client.ts
@@ -4,7 +4,7 @@ import {once} from "@e280/stz"
 import {Fns} from "../../core/types.js"
 import {defaults} from "../defaults.js"
 import {WsRemoteOptions} from "./types.js"
-import {endpoint} from "../../core/endpoint.js"
+import {makeEndpoint} from "../../core/endpoint.js"
 import {Messenger} from "../messenger/messenger.js"
 import {WebSocketConduit} from "../messenger/conduits/web-socket.js"
 import {waitForSocketOpen} from "../messenger/parts/wait-for-socket-open.js"
@@ -37,7 +37,7 @@ export async function wsClient<ServerFns extends Fns>(
 		conduit,
 		tap,
 		timeout,
-		getLocalEndpoint: (serverside) => endpoint({
+		getLocalEndpoint: (serverside) => makeEndpoint({
 			tap,
 			fns: rpc(serverside),
 		}),